Add files via upload
An inverted index project is an index data structure storing a mapping from content, such as words 
or numbers, to its locations in a database file, or in a document or a set of documents. The purpose 
of an inverted index is to allow fast full text searches, at a cost of increased processing when a 
document is added to the database. The inverted file may be the database file itself, rather than its 
index. It is the most popular data structure used in document retrieval systems, used on a large scale 
for example in search engines. The purpose of this project is to implement the inverted search using 
Hash Algorithms